[DI-2348] WIP: refactor postProcessOperationsWithModels p2
